Add render prop and remove elementRef prop
